+++ packages/moonstone/Picker/tests/Picker-specs.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {mount, shallow} from 'enzyme';
+import {mount} from 'enzyme';
 import {Picker, PickerBase} from '../Picker';
 
 describe('Picker Specs', () => {
@@ -31,13 +31,13 @@ describe('Picker Specs', () => {
 
 
 	it('should be disabled when empty', function () {
-		const picker = shallow(
+		const picker = mount(
 			<PickerBase>
 				{[]}
 			</PickerBase>
 		);
 
-		const actual = picker.find('SpottablePicker').last().prop('disabled');
+		const actual = picker.find('Picker').last().prop('disabled');
 		expect(actual).to.be.true();
 	});
 });