+++ packages/spotlight/Spottable/Spottable.js
@@ -6,7 +6,7 @@
  * @module spotlight/Spottable
  */
 
-import {forward, forwardWithPrevent, handle} from '@enact/core/handle';
+import {forward, forwardWithPrevent, handle, preventDefault, stop} from '@enact/core/handle';
 import hoc from '@enact/core/hoc';
 import {is} from '@enact/core/keymap';
 import React from 'react';
@@ -15,7 +15,7 @@ import PropTypes from 'prop-types';
 
 import {getContainersForNode} from '../src/container';
 import {hasPointerMoved} from '../src/pointer';
-import Spotlight from '../src/spotlight';
+import {getDirection, Spotlight} from '../src/spotlight';
 
 /**
  * The class name for spottable components. In general, you do not need to directly access this class
@@ -69,7 +69,19 @@ const defaultConfig = {
 	 * @public
 	 * @memberof spotlight/Spottable.Spottable.defaultConfig
 	 */
-	emulateMouse: true
+	emulateMouse: true,
+
+	/**
+	 * An array of numbers representing keyCodes that should trigger mouse event
+	 * emulation when `emulateMouse` is `true`. If a keyCode equals a directional
+	 * key, then default 5-way navigation will be prevented when that key is pressed.
+	 *
+	 * @type {Number[]}
+	 * @default [13, 16777221]
+	 * @public
+	 * @memberof spotlight/Spottable.Spottable.defaultConfig
+	 */
+	selectionKeys: [ENTER_KEY, REMOTE_OK_KEY]
 };
 
 /**
@@ -88,7 +100,7 @@ const defaultConfig = {
  * @returns {Function} Spottable
  */
 const Spottable = hoc(defaultConfig, (config, Wrapped) => {
-	const {emulateMouse} = config;
+	const {emulateMouse, selectionKeys} = config;
 
 	return class extends React.Component {
 		static displayName = 'Spottable'
@@ -234,17 +246,30 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 			}
 		}
 
-		shouldEmulateMouse = ({currentTarget, repeat, type, which}) => {
-			return emulateMouse && !repeat && (
+		shouldEmulateMouse = (ev) => {
+			if (!emulateMouse) {
+				return;
+			}
+
+			const {currentTarget, repeat, type, which} = ev;
+
+			const keyCode = selectionKeys.find((value) => (
 				// emulate mouse events for any remote okay button event
 				which === REMOTE_OK_KEY ||
-				// or a non-keypress enter event or any enter event on a non-keyboard accessible
+				// or a non-keypress selection event or any selection event on a non-keyboard accessible
 				// control
 				(
-					which === ENTER_KEY &&
+					which === value &&
 					(type !== 'keypress' || !isKeyboardAccessible(currentTarget))
 				)
-			);
+			));
+
+			if (getDirection(keyCode)) {
+				preventDefault(ev);
+				stop(ev);
+			}
+
+			return keyCode && !repeat;
 		}
 
 		forwardSpotlightEvents = (ev, {onSpotlightDown, onSpotlightLeft, onSpotlightRight, onSpotlightUp}) => {
@@ -263,9 +288,9 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 			return true;
 		}
 
-		handleSelect = (ev) => {
-			// Only apply accelerator if handling select
-			if ((ev.which === REMOTE_OK_KEY) || (ev.which === ENTER_KEY)) {
+		handleSelect = ({which}) => {
+			// Only apply accelerator if handling a selection key
+			if (selectionKeys.find((value) => which === value)) {
 				if (selectCancelled || (lastSelectTarget && lastSelectTarget !== this)) {
 					return false;
 				}
@@ -275,10 +300,15 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		forwardAndResetLastSelectTarget = (ev, props) => {
+			const {keyCode} = ev;
+			const key = selectionKeys.find((value) => keyCode === value);
 			const notPrevented = forwardWithPrevent('onKeyUp', ev, props);
 
-			// bail early for non-enter keyup to avoid clearing lastSelectTarget prematurely
-			if (!is('enter', ev.keyCode)) {
+			// bail early for non-selection keyup to avoid clearing lastSelectTarget prematurely
+			if (
+				(!is('enter', keyCode) && !key) ||
+				(!getDirection(keyCode) && !key)
+			) {
 				return notPrevented;
 			}